Moving the PosRgsMq file from pipelines directory to model directory.
