+++ src/main/java/com/google/cloud/genomics/dataflow/model/PosRgsMq.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.google.cloud.genomics.dataflow.pipelines;
+package com.google.cloud.genomics.dataflow.model;
 
 import com.google.api.client.json.GenericJson;
 import com.google.api.client.util.Key;
